Support Dev Proxy install from package managers. Closes #82
Closes #82
