+++ CHANGELOG.md
@@ -5,9 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2024-05-30
+
+### Changed:
+
+- Notification: Support for installing Dev Proxy via package manager when not installed
+
 ## [0.3.1] - 2024-05-22
 
-### Fixed
+### Fixed:
 
 - Fixed issue where `configSection` diagnostic check required `RateLimitingPlugin` to have a `configSection`
