+++ package.json
@@ -2,7 +2,7 @@
   "name": "dev-proxy-toolkit",
   "displayName": "Dev Proxy Toolkit",
   "description": "Makes it easy to create and update Dev Proxy configuration files.",
-  "version": "0.3.1",
+  "version": "0.4.0",
   "publisher": "garrytrinder",
   "engines": {
     "vscode": "^1.85.0"
